fix: Cleaned up code structure
