+++ .storybook/config.js
@@ -1,10 +1,5 @@
 import React from 'react';
 import { configure, addDecorator } from '@storybook/react';
-import { BlueRainDecorator } from '../storybook/bluerain';
-
-// Add BlueRain
-const BRConfigs = require('../bluerain');
-addDecorator(BlueRainDecorator(BRConfigs));
 
 // automatically import all files ending in *.stories.js
 const req = require.context('../src', true, /.stories.tsx$/);
