+++ package.json
@@ -26,7 +26,7 @@
   "scripts": {
     "lint": "tslint -t codeFrame 'src/**/*.ts' 'test/**/*.ts'",
     "prebuild": "rimraf dist",
-    "build": "rimraf dist && bluerain build plugin",
+    "build": "tsc && bluerain build plugin",
     "start": "tsc -w",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -69,7 +69,7 @@
         "./src"
       ],
       "pattern": "**/*.stories.tsx",
-      "outputFile": "./storybook/storyLoader.tsx"
+      "outputFile": "./storybook/storyLoader.js"
     }
   },
   "jest": {
@@ -102,6 +102,7 @@
     "@blueeast/bluerain-plugin-react-router": "^1.1.1",
     "@blueeast/bluerain-plugin-redux": "^0.6.2",
     "@blueeast/bluerain-plugin-responsive-components": "^1.1.3",
+    "@blueeast/bluerain-storybook-addon": "^0.1.0",
     "@blueeast/tslint-config-blueeast": "^0.4.0",
     "@storybook/addon-actions": "^3.3.12",
     "@storybook/addon-graphql": "^3.2.0",
@@ -119,7 +120,6 @@
     "awesome-typescript-loader": "^3.4.1",
     "babel-plugin-react-native-web": "^0.5.2",
     "babel-preset-react-native": "^4.0.0",
-    "bluerain-storybook-addon": "^1.0.0",
     "bootstrap-loader": "^2.0.0-beta.16",
     "bundlesize": "^0.15.2",
     "codecov": "^2.3.1",
