+++ src/FA/component/index.android.ts
@@ -1,0 +1,2 @@
+import { FontAwesome } from '@expo/vector-icons';
+export default FontAwesome;
