+++ src/FA/component/index.ios.ts
@@ -1,0 +1,2 @@
+import { FontAwesome } from '@expo/vector-icons';
+export default FontAwesome;
