+++ src/FA/component/index.ts
@@ -1,0 +1,2 @@
+import Icon from 'react-native-vector-icons/dist/FontAwesome';
+export default Icon;
