+++ src/FA/setup/index.ts
@@ -1,0 +1,3 @@
+import { setup as genericSetup } from '../../setup';
+
+genericSetup(require('react-native-vector-icons/Fonts/FontAwesome.ttf'));
