--- src/FontAwesomeIcons/index.android.tsx
@@ -1,4 +1,0 @@
-
-import VectorFontAwesome from  './bluerain-plugin-vector-iconsFontAwesome';
-
-export {VectorFontAwesome};
