--- src/MaterialIcons/index.ios.tsx
@@ -1,4 +1,0 @@
-
-import VectorMaterialIcons from  './bluerain-plugin-vector-icons-MaterialIcons';
-
-export {VectorMaterialIcons};
