--- src/VectorIonIcons/index.android.tsx
@@ -1,4 +1,0 @@
-
-import VectorIonicons from './bluerain-plugin-vector-icons-Ionicons';
-
-export {VectorIonicons};
