+++ src/setup/index.android.ts
@@ -1,0 +1,1 @@
+export const setup = iconPath => {};
