+++ src/setup/index.ios.ts
@@ -1,0 +1,1 @@
+export const setup = iconPath => {};
