--- storybook/bluerain/index.ts
@@ -1,4 +1,0 @@
-import BlueRainDecorator from './BlueRainDecorator';
-import bluerainStory from './BlueRainStory';
-
-export { BlueRainDecorator, bluerainStory };
