--- storybook/storiesOf/index.ios.js
@@ -1,2 +1,0 @@
-import { storiesOf } from '@storybook/react-native';
-export default storiesOf;
