--- storybook/storiesOf/index.ts
@@ -1,2 +1,0 @@
-import { storiesOf } from '@storybook/react';
-export default storiesOf;
