--- typings/index.d.ts
@@ -1,4 +1,0 @@
-declare module "*.ttf" {
-	const value: any;
-	export default value;
-}
